add context to liquibase, disable loading initial data into test

+++ ski_jumping.domain/src/main/java/com/pl/skijumping/domain/repository/CountryRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.querydsl.QueryDslPredicateExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CountryRepository extends JpaRepository<Country, Long>, QueryDslPredicateExecutor {
+public interface CountryRepository extends JpaRepository<Country, Integer>, QueryDslPredicateExecutor {
 }
